Lanjut Fitur Manajemen Siswa

+++ database/migrations/2020_03_12_043825_create_siswas_table.php
@@ -19,7 +19,7 @@ class CreateSiswasTable extends Migration
             $table->string('pendidikan_terakhir');
             $table->integer('umur');
             $table->text('alamat_lengkap');
-            $table->integer('no_telp');
+            $table->bigInteger('no_telp');
             $table->string('email')->nullable();
             $table->timestamps();
             $table->softDeletes();
